+++ src/actions/index.js
@@ -1,0 +1,7 @@
+import apartmentsActions from './apartments'
+
+const allActions = {
+	apartmentsActions
+};
+
+export default allActions;