+++ src/components/App/index.js
@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
+import logo from '../../logo.svg';
+import './index.css';
+import Button from '../../components/Button';
+import Results from '../../components/Results';
 
-class App extends Component {
+class Index extends Component {
   render() {
     return (
       <div className="App">
@@ -11,11 +13,11 @@ class App extends Component {
           <h2>Welcome to React</h2>
         </div>
         <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
+          <Results/>
         </p>
       </div>
     );
   }
 }
 
-export default App;
+export default Index;