+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import apartments from './apartmentsReducer'
+import {combineReducers} from 'redux'
+
+const rootReducer = combineReducers({
+	apartments
+});
+
+export default rootReducer;